* wxWidget's wizard: Changed to support the wxTeam made Binaries (Thanks stahta01).
Added function SilentVerifyWxWinLibFile that checks for binaries in wx.lib folder.
Added code that results in wx.lib and wx.include being in CB project.
Added call to GetCompilerIncludeDir and VerifyFile.
This resulted in a different message if headers are missing.
This commit also contains the removal of GetUserVariableManager().Exists() call.